Consolidate Cagra index merge tests into CagraBuildAndSearchIT; resolve whitespace and formatting issues
