+++ java/cuvs-java/src/test/java/com/nvidia/cuvs/CagraBuildAndSearchIT.java
@@ -18,6 +18,7 @@ package com.nvidia.cuvs;
 
 import static com.carrotsearch.randomizedtesting.RandomizedTest.assumeTrue;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -42,6 +43,7 @@ import org.slf4j.LoggerFactory;
 import com.carrotsearch.randomizedtesting.RandomizedRunner;
 import com.nvidia.cuvs.CagraIndexParams.CagraGraphBuildAlgo;
 import com.nvidia.cuvs.CagraIndexParams.CuvsDistanceType;
+import com.nvidia.cuvs.CagraMergeParams.MergeStrategy;
 
 @RunWith(RandomizedRunner.class)
 public class CagraBuildAndSearchIT extends CuVSTestCase {
@@ -195,4 +197,231 @@ public class CagraBuildAndSearchIT extends CuVSTestCase {
     };
     return thread;
   }
+
+  @Test
+    public void testMergingIndexes() throws Throwable {
+        float[][] dataset1 = {
+            {0.0f, 0.0f},
+            {1.0f, 1.0f}
+        };
+
+        float[][] dataset2 = {
+            {10.0f, 10.0f},
+            {11.0f, 11.0f}
+        };
+
+        float[][] queries = {
+            {1.0f, 1.0f},    // Should be closest to dataset1[1] -> index 1
+            {10.5f, 10.5f},  // Should be closest to dataset2[0] -> index 2
+            {0.0f, 0.0f}     // Should be closest to dataset1[0] -> index 0
+        };
+
+        // Expected search results for each query (nearest neighbor and its distance)
+        List<Map<Integer, Float>> expectedResults = Arrays.asList(
+            Map.of(1, 0.0f, 0, 2.0f, 2, 162.0f),
+            Map.of(2, 0.5f, 3, 0.5f, 1, 180.5f),
+            Map.of(0, 0.0f, 1, 2.0f, 2, 200.0f)
+        );
+
+        try (CuVSResources resources = CuVSResources.create()) {
+            CagraIndexParams indexParams = new CagraIndexParams.Builder()
+                .withCagraGraphBuildAlgo(CagraGraphBuildAlgo.NN_DESCENT)
+                .withGraphDegree(1)
+                .withIntermediateGraphDegree(2)
+                .withNumWriterThreads(4)
+                .withMetric(CuvsDistanceType.L2Expanded)
+                .build();
+
+            log.info("Building first index...");
+            CagraIndex index1 = CagraIndex.newBuilder(resources)
+                .withDataset(dataset1)
+                .withIndexParams(indexParams)
+                .build();
+
+            log.info("Building second index...");
+            CagraIndex index2 = CagraIndex.newBuilder(resources)
+                .withDataset(dataset2)
+                .withIndexParams(indexParams)
+                .build();
+
+            log.info("Merging indexes...");
+            CagraIndex mergedIndex = CagraIndex.merge(new CagraIndex[]{index1, index2});
+            log.info("Merge completed successfully");
+
+            CagraSearchParams searchParams = new CagraSearchParams.Builder(resources)
+                .build();
+
+            List<Integer> mergedMap = Arrays.asList(0, 1, 2, 3);
+            CagraQuery query = new CagraQuery.Builder()
+                .withTopK(3)
+                .withSearchParams(searchParams)
+                .withQueryVectors(queries)
+                .withMapping(mergedMap)
+                .build();
+
+                log.info("Searching merged index...");
+                SearchResults results = mergedIndex.search(query);
+                log.info("Search results: " + results.getResults().toString());
+
+                assertEquals(expectedResults, results.getResults());
+
+                // --- Serialization/deserialization check ---
+                String indexFileName = java.util.UUID.randomUUID().toString() + ".cag";
+                mergedIndex.serialize(new java.io.FileOutputStream(indexFileName));
+
+                java.io.File indexFile = new java.io.File(indexFileName);
+                java.io.InputStream inputStream = new java.io.FileInputStream(indexFile);
+                CagraIndex loadedMergedIndex = CagraIndex.newBuilder(resources)
+                    .from(inputStream)
+                    .build();
+
+                SearchResults resultsFromLoaded = loadedMergedIndex.search(query);
+                assertEquals(expectedResults, resultsFromLoaded.getResults());
+
+                if (indexFile.exists()) {
+                    indexFile.delete();
+                }
+                index1.destroyIndex();
+                index2.destroyIndex();
+                mergedIndex.destroyIndex();
+                loadedMergedIndex.destroyIndex();
+
+        }
+    }
+
+    @Test
+    public void testMergeStrategies() throws Throwable {
+        float[][] dataset1 = {
+            {0.0f, 0.0f},
+            {1.0f, 1.0f}
+        };
+
+        float[][] dataset2 = {
+            {10.0f, 10.0f},
+            {11.0f, 11.0f}
+        };
+
+        float[][] queries = {
+            {1.0f, 1.0f},
+            {10.5f, 10.5f},
+            {0.0f, 0.0f}
+        };
+
+        List<Map<Integer, Float>> expectedResults = Arrays.asList(
+            Map.of(1, 0.0f, 0, 2.0f, 2, 162.0f),
+            Map.of(2, 0.5f, 3, 0.5f, 1, 180.5f),
+            Map.of(0, 0.0f, 1, 2.0f, 2, 200.0f)
+        );
+
+        try (CuVSResources resources = CuVSResources.create()) {
+            CagraIndexParams indexParams = new CagraIndexParams.Builder()
+                .withCagraGraphBuildAlgo(CagraGraphBuildAlgo.NN_DESCENT)
+                .withGraphDegree(1)
+                .withIntermediateGraphDegree(2)
+                .withNumWriterThreads(4)
+                .withMetric(CuvsDistanceType.L2Expanded)
+                .build();
+
+            log.info("Building first index...");
+            CagraIndex index1 = CagraIndex.newBuilder(resources)
+                .withDataset(dataset1)
+                .withIndexParams(indexParams)
+                .build();
+
+            log.info("Building second index...");
+            CagraIndex index2 = CagraIndex.newBuilder(resources)
+                .withDataset(dataset2)
+                .withIndexParams(indexParams)
+                .build();
+
+            CagraIndexParams outputIndexParams = new CagraIndexParams.Builder()
+                .withCagraGraphBuildAlgo(CagraGraphBuildAlgo.NN_DESCENT)
+                .withGraphDegree(2)
+                .withIntermediateGraphDegree(4)
+                .withNumWriterThreads(4)
+                .withMetric(CuvsDistanceType.L2Expanded)
+                .build();
+
+            CagraMergeParams physicalMergeParams = new CagraMergeParams.Builder()
+                .withOutputIndexParams(outputIndexParams)
+                .withStrategy(MergeStrategy.PHYSICAL)
+                .build();
+
+            log.info("Merging indexes with PHYSICAL strategy...");
+            CagraIndex physicalMergedIndex = CagraIndex.merge(new CagraIndex[]{index1, index2}, physicalMergeParams);
+            log.info("Physical merge completed successfully");
+
+            CagraMergeParams logicalMergeParams = new CagraMergeParams.Builder()
+                .withOutputIndexParams(outputIndexParams)
+                .withStrategy(MergeStrategy.LOGICAL)
+                .build();
+
+            log.info("Merging indexes with LOGICAL strategy...");
+            CagraIndex logicalMergedIndex = CagraIndex.merge(new CagraIndex[]{index1, index2}, logicalMergeParams);
+            log.info("Logical merge completed successfully");
+
+            CagraSearchParams searchParams = new CagraSearchParams.Builder(resources)
+                .build();
+
+            List<Integer> mergedMap = Arrays.asList(0, 1, 2, 3);
+            CagraQuery query = new CagraQuery.Builder()
+                .withTopK(3)
+                .withSearchParams(searchParams)
+                .withQueryVectors(queries)
+                .withMapping(mergedMap)
+                .build();
+
+            log.info("Searching physically merged index...");
+            SearchResults physicalResults = physicalMergedIndex.search(query);
+            assertNotNull("Physical merge search results should not be null", physicalResults);
+            assertEquals("Physical merge search results should match expected", expectedResults, physicalResults.getResults());
+
+            log.info("Searching logically merged index...");
+            SearchResults logicalResults = logicalMergedIndex.search(query);
+            assertNotNull("Logical merge search results should not be null", logicalResults);
+            assertEquals("Logical merge search results should match expected", expectedResults, logicalResults.getResults());
+
+            // --- Serialization/deserialization check for both merged indexes ---
+            String physicalIndexFileName = java.util.UUID.randomUUID().toString() + ".cag";
+            physicalMergedIndex.serialize(new java.io.FileOutputStream(physicalIndexFileName));
+
+            String logicalIndexFileName = java.util.UUID.randomUUID().toString() + ".cag";
+            logicalMergedIndex.serialize(new java.io.FileOutputStream(logicalIndexFileName));
+
+            java.io.File physicalIndexFile = new java.io.File(physicalIndexFileName);
+            java.io.InputStream physicalInputStream = new java.io.FileInputStream(physicalIndexFile);
+            CagraIndex loadedPhysicalIndex = CagraIndex.newBuilder(resources)
+                .from(physicalInputStream)
+                .build();
+
+            java.io.File logicalIndexFile = new java.io.File(logicalIndexFileName);
+            java.io.InputStream logicalInputStream = new java.io.FileInputStream(logicalIndexFile);
+            CagraIndex loadedLogicalIndex = CagraIndex.newBuilder(resources)
+                .from(logicalInputStream)
+                .build();
+
+            SearchResults resultsFromLoadedPhysical = loadedPhysicalIndex.search(query);
+            assertEquals("Loaded physical index search results should match expected",
+                expectedResults, resultsFromLoadedPhysical.getResults());
+
+            SearchResults resultsFromLoadedLogical = loadedLogicalIndex.search(query);
+            assertEquals("Loaded logical index search results should match expected",
+                expectedResults, resultsFromLoadedLogical.getResults());
+
+            if (physicalIndexFile.exists()) {
+                physicalIndexFile.delete();
+            }
+            if (logicalIndexFile.exists()) {
+                logicalIndexFile.delete();
+            }
+
+            index1.destroyIndex();
+            index2.destroyIndex();
+            physicalMergedIndex.destroyIndex();
+            logicalMergedIndex.destroyIndex();
+            loadedPhysicalIndex.destroyIndex();
+            loadedLogicalIndex.destroyIndex();
+        }
+    }
+
 }
